14. meta-ad-collector/requirements.txt
requests==2.31.0
pandas==2.1.4
gspread==6.0.2
google-auth==2.27.0
playwright==1.41.0
python-dotenv==1.0.0
schedule==1.2.1
rich==13.7.0